Dbcontext Name rename and UnitOfWork name reaname all is done and working.

+++ DevSlkill.Inventory/DevSkill.Inventory.Domain/Entities/Product.cs
@@ -11,8 +11,8 @@ namespace DevSkill.Inventory.Domain.Entities
         public Guid Id { get; set; }
         public string ProductName { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty ;
-        public decimal Price { get; set; }
-        public decimal Ratings { get; set; }
+        public int Price { get; set; }
+        public int Ratings { get; set; }
 
         public Product()
         {
