+++ DevSlkill.Inventory/DevSkill.Inventory.Web/Areas/Admin/Models/ProductModel.cs
@@ -7,7 +7,7 @@ namespace DevSkill.Inventory.Web.Areas.Admin.Models
         public Guid Id { get; set; }
         public string ProductName { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public decimal Price { get; set; }
-        public decimal Ratings { get; set; }
+        public int Price { get; set; }
+        public int Ratings { get; set; }
     }
 }
